The JSON parser is ready to go into Beta testing.
Use importJS(jsobject) to load documents.  The JSON Parser creates an
XML document, and then passes it to the XML parser.  The JSON is not
checked for accurracy or completeness, and relies on the XML parser for
that.

CreateX3DFromString needs to be tested with JSON strings.  I will work
on that.  We can even take that part out, since I am not using it
currently.

The JSON Parser is a subset of X3DJSONLD.js.  The website is
https://coderextreme.net/X3DJSONLD.

JSON Documents are Available in X3D Resources Examples, or with the
X3dToJson.xslt stylesheet to convert from XML.